to: Normalize file paths in code search query
to: remove unuse import
